+++ src/lib/project/ejectEngineChangelog.ts
@@ -94,6 +94,24 @@ const changes: EngineChange[] = [
       "src/core/vm_ui.c",
     ],
   },
+  {
+    version: "3.0.3-e1",
+    description:
+      "Updates:\n" +
+      "   * Avoid rendering garbage when no scene has loaded yet" +
+      "   * use GBDK-2020 hardware sprite hiding function" +
+      "   * Move save blob signature to game data",
+    modifiedFiles: [
+      "include/actor.h",
+      "include/oam_utils.h",
+      "include/shadow.h",
+      "src/core/actor.c",
+      "src/core/data_manager.c",
+      "src/core/interrupts.c",
+      "src/core/load_save.c",
+      "src/core/oam_utils.s",
+    ],
+  },
 ];
 
 const ejectEngineChangelog = (currentVersion: string) => {